requirements.txt
Django==3.2.12
djangorestframework==3.12.4
